requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
pinecone-client==2.2.4
langchain==0.1.16
langchain-community
flask==2.3.2
pypdf
python-dotenv




-e .